python_api/requirements_render.txt
# Core FastAPI and web server dependencies
fastapi==0.110.1
uvicorn==0.25.0
python-multipart==0.0.20

# Database and Supabase
supabase==2.22.0
postgrest==2.22.0
realtime==2.22.0
storage3==2.22.0

# Authentication and security
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.1.3
PyJWT==2.10.1
cryptography==46.0.2

# Email and communication
aiosmtplib==4.0.2
requests==2.32.5
httpx==0.28.1

# Data handling
pydantic==2.12.0
pydantic_core==2.41.1
python-dateutil==2.9.0.post0

# Environment and configuration
python-dotenv==1.1.1

# Utilities
typing_extensions==4.15.0
annotated-types==0.7.0
